backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
drf-spectacular==0.28.0
django-allauth==65.12.1
PyJWT==2.10.1
PyPDFForm==3.5.4
pdfplumber==0.11.7
pytesseract==0.3.13
pdf2image==1.17.0
openai==2.5.0
python-dotenv==1.1.1
Pillow==11.2.0
openpyxl==3.1.5
WeasyPrint==63.1
gunicorn==23.0.0

